Update opsman VM OAuth scopes scopes after we create it
Platform Automation's create-vm task creates the Opsman VM without the necessary Compute Engine OAuth scopes to perform the validation checks on tile configuration (even though the service account we provide in opsman.yml has a role with sufficient permissions - even setting that role to project editor doesn't fix this). As a result, when you try to save such tile configuration you get an 'IaasConfigurationVerifier' error. To get round this we have to stop the VM, update its scopes, and then start it again and decrypt it.